# Optimization: make the most of Hint::AcceptsSingular when call make_scalar_function to Improve performance