--- a/datafusion/functions/benches/overlay.rs
+++ b/datafusion/functions/benches/overlay.rs
@@ -15,15 +15,18 @@
 // specific language governing permissions and limitations
 // under the License.
 
-use arrow::array::StringArray;
+use arrow::array::{Array, Int64Array, StringArray};
 use criterion::{criterion_group, criterion_main, BenchmarkId, Criterion};
 use datafusion_common::ScalarValue;
 use datafusion_expr::ColumnarValue;
 use datafusion_functions::string;
 use std::sync::Arc;
 
-fn create_4args(size: usize) -> Vec<ColumnarValue> {
+/// Create four args, three of which are Scalars and one is a StringArray.
+/// The `size` represents the length of the StringArray.
+fn create_4args_with_3scalars(size: usize) -> Vec<ColumnarValue> {
     let array: StringArray = std::iter::repeat(Some("Txxxxas")).take(size).collect();
+    assert_eq!(array.len(), size);
     let characters = ScalarValue::Utf8(Some("hom".to_string()));
     let pos = ScalarValue::Int64(Some(2));
     let len = ScalarValue::Int64(Some(4));
@@ -35,11 +38,37 @@ fn create_4args(size: usize) -> Vec<ColumnarValue> {
     ]
 }
 
+/// Create four args, all of which are Arrays.
+/// The `size` represents the length of Array.
+fn create_4args_without_scalar(size: usize) -> Vec<ColumnarValue> {
+    let array: StringArray = std::iter::repeat(Some("Txxxxas")).take(size).collect();
+    let characters: StringArray = std::iter::repeat(Some("hom")).take(size).collect();
+    let pos: Int64Array = std::iter::repeat(Some(2)).take(size).collect();
+    let len: Int64Array = std::iter::repeat(Some(4)).take(size).collect();
+    vec![
+        ColumnarValue::Array(Arc::new(array)),
+        ColumnarValue::Array(Arc::new(characters)),
+        ColumnarValue::Array(Arc::new(pos)),
+        ColumnarValue::Array(Arc::new(len)),
+    ]
+}
+
 fn criterion_benchmark(c: &mut Criterion) {
     let overlay = string::overlay();
-    for size in [1024, 4096, 8192] {
-        let args = create_4args(size);
-        let mut group = c.benchmark_group("overlay_with_4args");
+    let sizes: Vec<usize> = vec![1024, 4096, 8192];
+
+    for size in &sizes {
+        let args = create_4args_with_3scalars(*size);
+        let mut group = c.benchmark_group("4args_with_3scalars");
+        group.bench_function(BenchmarkId::new("overlay", size), |b| {
+            b.iter(|| criterion::black_box(overlay.invoke(&args).unwrap()))
+        });
+        group.finish();
+    }
+
+    for size in &sizes {
+        let args = create_4args_without_scalar(*size);
+        let mut group = c.benchmark_group("4args_without_scalar");
         group.bench_function(BenchmarkId::new("overlay", size), |b| {
             b.iter(|| criterion::black_box(overlay.invoke(&args).unwrap()))
         });