--- a/datafusion/functions/src/string/overlay.rs
+++ b/datafusion/functions/src/string/overlay.rs
@@ -18,7 +18,7 @@
 use std::any::Any;
 use std::sync::Arc;
 
-use arrow::array::{ArrayRef, GenericStringArray, OffsetSizeTrait};
+use arrow::array::{Array, ArrayRef, GenericStringArray, OffsetSizeTrait};
 use arrow::datatypes::DataType;
 
 use datafusion_common::cast::{as_generic_string_array, as_int64_array};
@@ -28,7 +28,7 @@ use datafusion_expr::{ColumnarValue, Volatility};
 use datafusion_expr::{ScalarUDFImpl, Signature};
 use datafusion_physical_expr::functions::Hint;
 
-use crate::utils::{adaptive_array_iter, make_scalar_function, utf8_to_str_type};
+use crate::utils::{make_scalar_function, utf8_to_str_type, ScalarFlags};
 
 #[derive(Debug)]
 pub struct OverlayFunc {
@@ -94,6 +94,80 @@ impl ScalarUDFImpl for OverlayFunc {
     }
 }
 
+pub fn overlay3<'a, T: OffsetSizeTrait>(
+    string_array_iter: impl Iterator<Item = Option<&'a str>>,
+    characters_array_iter: impl Iterator<Item = Option<&'a str>>,
+    pos_num_iter: impl Iterator<Item = Option<i64>>,
+) -> Result<ArrayRef> {
+    let result = string_array_iter
+        .zip(characters_array_iter)
+        .zip(pos_num_iter)
+        .map(|((string, characters), start_pos)| {
+            match (string, characters, start_pos) {
+                (Some(string), Some(characters), Some(start_pos)) => {
+                    let string_len = string.chars().count();
+                    let characters_len = characters.chars().count();
+                    let replace_len = characters_len as i64;
+                    let mut res = String::with_capacity(string_len.max(characters_len));
+
+                    //as sql replace index start from 1 while string index start from 0
+                    if start_pos > 1 && start_pos - 1 < string_len as i64 {
+                        let start = (start_pos - 1) as usize;
+                        res.push_str(&string[..start]);
+                    }
+                    res.push_str(characters);
+                    // if start + replace_len - 1 >= string_length, just to string end
+                    if start_pos + replace_len - 1 < string_len as i64 {
+                        let end = (start_pos + replace_len - 1) as usize;
+                        res.push_str(&string[end..]);
+                    }
+                    Ok(Some(res))
+                }
+                _ => Ok(None),
+            }
+        })
+        .collect::<Result<GenericStringArray<T>>>()?;
+    Ok(Arc::new(result) as ArrayRef)
+}
+
+pub fn overlay4<'a, T: OffsetSizeTrait>(
+    string_array_iter: impl Iterator<Item = Option<&'a str>>,
+    characters_array_iter: impl Iterator<Item = Option<&'a str>>,
+    pos_num_iter: impl Iterator<Item = Option<i64>>,
+    len_num_iter: impl Iterator<Item = Option<i64>>,
+) -> Result<ArrayRef> {
+    let result = string_array_iter
+        .zip(characters_array_iter)
+        .zip(pos_num_iter)
+        .zip(len_num_iter)
+        .map(|(((string, characters), start_pos), len)| {
+            match (string, characters, start_pos, len) {
+                (Some(string), Some(characters), Some(start_pos), Some(len)) => {
+                    let string_len = string.chars().count();
+                    let characters_len = characters.chars().count();
+                    let replace_len = len.min(string_len as i64);
+                    let mut res = String::with_capacity(string_len.max(characters_len));
+
+                    //as sql replace index start from 1 while string index start from 0
+                    if start_pos > 1 && start_pos - 1 < string_len as i64 {
+                        let start = (start_pos - 1) as usize;
+                        res.push_str(&string[..start]);
+                    }
+                    res.push_str(characters);
+                    // if start + replace_len - 1 >= string_length, just to string end
+                    if start_pos + replace_len - 1 < string_len as i64 {
+                        let end = (start_pos + replace_len - 1) as usize;
+                        res.push_str(&string[end..]);
+                    }
+                    Ok(Some(res))
+                }
+                _ => Ok(None),
+            }
+        })
+        .collect::<Result<GenericStringArray<T>>>()?;
+    Ok(Arc::new(result) as ArrayRef)
+}
+
 /// OVERLAY(string1 PLACING string2 FROM integer FOR integer2)
 /// Replaces a substring of string1 with string2 starting at the integer bit
 /// pgsql overlay('Txxxxas' placing 'hom' from 2 for 4) → Thomas
@@ -104,84 +178,31 @@ pub fn overlay<T: OffsetSizeTrait>(args: &[ArrayRef]) -> Result<ArrayRef> {
             let string_array = as_generic_string_array::<T>(&args[0])?;
             let characters_array = as_generic_string_array::<T>(&args[1])?;
             let pos_num = as_int64_array(&args[2])?;
+            let flag = ScalarFlags::try_create(args)?;
 
-            let characters_array_iter = adaptive_array_iter(characters_array.iter());
-            let pos_num_iter = adaptive_array_iter(pos_num.iter());
-
-            let result = string_array
-                .iter()
-                .zip(characters_array_iter)
-                .zip(pos_num_iter)
-                .map(|((string, characters), start_pos)| {
-                    match (string, characters, start_pos) {
-                        (Some(string), Some(characters), Some(start_pos)) => {
-                            let string_len = string.chars().count();
-                            let characters_len = characters.chars().count();
-                            let replace_len = characters_len as i64;
-                            let mut res =
-                                String::with_capacity(string_len.max(characters_len));
-
-                            //as sql replace index start from 1 while string index start from 0
-                            if start_pos > 1 && start_pos - 1 < string_len as i64 {
-                                let start = (start_pos - 1) as usize;
-                                res.push_str(&string[..start]);
-                            }
-                            res.push_str(characters);
-                            // if start + replace_len - 1 >= string_length, just to string end
-                            if start_pos + replace_len - 1 < string_len as i64 {
-                                let end = (start_pos + replace_len - 1) as usize;
-                                res.push_str(&string[end..]);
-                            }
-                            Ok(Some(res))
-                        }
-                        _ => Ok(None),
-                    }
-                })
-                .collect::<Result<GenericStringArray<T>>>()?;
-            Ok(Arc::new(result) as ArrayRef)
+            invoke_function!(
+                flag,
+                |arg0, arg1, arg2| overlay3::<T>(arg0, arg1, arg2),
+                string_array,
+                characters_array,
+                pos_num
+            )
         }
         4 => {
             let string_array = as_generic_string_array::<T>(&args[0])?;
             let characters_array = as_generic_string_array::<T>(&args[1])?;
             let pos_num = as_int64_array(&args[2])?;
             let len_num = as_int64_array(&args[3])?;
+            let flag = ScalarFlags::try_create(args)?;
 
-            let characters_array_iter = adaptive_array_iter(characters_array.iter());
-            let pos_num_iter = adaptive_array_iter(pos_num.iter());
-            let len_num_iter = adaptive_array_iter(len_num.iter());
-
-            let result = string_array
-                .iter()
-                .zip(characters_array_iter)
-                .zip(pos_num_iter)
-                .zip(len_num_iter)
-                .map(|(((string, characters), start_pos), len)| {
-                    match (string, characters, start_pos, len) {
-                        (Some(string), Some(characters), Some(start_pos), Some(len)) => {
-                            let string_len = string.chars().count();
-                            let characters_len = characters.chars().count();
-                            let replace_len = len.min(string_len as i64);
-                            let mut res =
-                                String::with_capacity(string_len.max(characters_len));
-
-                            //as sql replace index start from 1 while string index start from 0
-                            if start_pos > 1 && start_pos - 1 < string_len as i64 {
-                                let start = (start_pos - 1) as usize;
-                                res.push_str(&string[..start]);
-                            }
-                            res.push_str(characters);
-                            // if start + replace_len - 1 >= string_length, just to string end
-                            if start_pos + replace_len - 1 < string_len as i64 {
-                                let end = (start_pos + replace_len - 1) as usize;
-                                res.push_str(&string[end..]);
-                            }
-                            Ok(Some(res))
-                        }
-                        _ => Ok(None),
-                    }
-                })
-                .collect::<Result<GenericStringArray<T>>>()?;
-            Ok(Arc::new(result) as ArrayRef)
+            invoke_function!(
+                flag,
+                |arg0, arg1, arg2, arg3| overlay4::<T>(arg0, arg1, arg2, arg3),
+                string_array,
+                characters_array,
+                pos_num,
+                len_num
+            )
         }
         other => {
             exec_err!("overlay was called with {other} arguments. It requires 3 or 4.")